+++ app/src/main/java/com/skgtecnologia/sisem/data/auth/AuthRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.skgtecnologia.sisem.data.auth
 import com.skgtecnologia.sisem.data.auth.cache.AuthCacheDataSource
 import com.skgtecnologia.sisem.data.auth.remote.AuthRemoteDataSource
 import com.skgtecnologia.sisem.domain.auth.AuthRepository
-import com.skgtecnologia.sisem.domain.login.model.AccessTokenModel
+import com.skgtecnologia.sisem.domain.auth.model.AccessTokenModel
 import javax.inject.Inject
 
 class AuthRepositoryImpl @Inject constructor(
@@ -15,18 +15,16 @@ class AuthRepositoryImpl @Inject constructor(
         authRemoteDataSource.authenticate(
             username = username,
             password = password,
-            turnId = authCacheDataSource.retrieveAccessToken()?.turn?.id?.toString() ?: ""
-        )
-            .onSuccess { accessTokenModel ->
-                authCacheDataSource.storeAccessToken(accessTokenModel)
-            }
-            .getOrThrow()
+            turnId = authCacheDataSource.retrieveAccessToken()?.turn?.id?.toString().orEmpty()
+        ).onSuccess { accessTokenModel ->
+            authCacheDataSource.storeAccessToken(accessTokenModel)
+        }.getOrThrow()
 
     override suspend fun getAccessToken(): String? =
         authCacheDataSource.retrieveAccessToken()?.accessToken
 
-    override suspend fun getAllAccessToken(): List<AccessTokenModel> =
-        authCacheDataSource.getAllAccessTokenModel()
+    override suspend fun getAllAccessTokens(): List<AccessTokenModel> =
+        authCacheDataSource.retrieveAllAccessTokens()
 
     override suspend fun logout(username: String): String =
         authRemoteDataSource.logout(username).getOrThrow()
