+++ app/src/main/java/com/skgtecnologia/sisem/data/auth/cache/AuthCacheDataSource.kt
@@ -4,7 +4,7 @@ import androidx.annotation.CheckResult
 import com.skgtecnologia.sisem.data.auth.cache.dao.AccessTokenDao
 import com.skgtecnologia.sisem.data.auth.cache.model.mapToCache
 import com.skgtecnologia.sisem.data.auth.cache.model.mapToDomain
-import com.skgtecnologia.sisem.domain.login.model.AccessTokenModel
+import com.skgtecnologia.sisem.domain.auth.model.AccessTokenModel
 import javax.inject.Inject
 
 class AuthCacheDataSource @Inject constructor(
@@ -20,6 +20,6 @@ class AuthCacheDataSource @Inject constructor(
         accessTokenDao.getAccessToken()?.mapToDomain()
 
     @CheckResult
-    suspend fun getAllAccessTokenModel(): List<AccessTokenModel> =
-        accessTokenDao.getAllAccessToken().map { it.mapToDomain() }
+    suspend fun retrieveAllAccessTokens(): List<AccessTokenModel> =
+        accessTokenDao.getAllAccessTokens().map { it.mapToDomain() }
 }
