+++ app/src/main/java/com/skgtecnologia/sisem/data/auth/cache/dao/AccessTokenDao.kt
@@ -12,9 +12,9 @@ interface AccessTokenDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertAccessToken(accessTokenEntity: AccessTokenEntity)
 
-    @Query("SELECT * FROM access_token LIMIT 1") // FIXME: Adjust this query by Role
+    @Query("SELECT * FROM access_token LIMIT 1")
     suspend fun getAccessToken(): AccessTokenEntity?
 
     @Query("SELECT * FROM access_token")
-    suspend fun getAllAccessToken(): List<AccessTokenEntity>
+    suspend fun getAllAccessTokens(): List<AccessTokenEntity>
 }
