+++ app/src/main/java/com/skgtecnologia/sisem/data/auth/cache/model/PreOperationalEntity.kt
@@ -3,7 +3,7 @@ package com.skgtecnologia.sisem.data.auth.cache.model
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.skgtecnologia.sisem.domain.login.model.PreOperationalModel
+import com.skgtecnologia.sisem.domain.auth.model.PreOperationalModel
 
 @Entity(tableName = "pre_operational",)
 data class PreOperationalEntity(
