+++ app/src/main/java/com/skgtecnologia/sisem/data/auth/cache/model/TurnEntity.kt
@@ -3,7 +3,7 @@ package com.skgtecnologia.sisem.data.auth.cache.model
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.skgtecnologia.sisem.domain.login.model.TurnModel
+import com.skgtecnologia.sisem.domain.auth.model.TurnModel
 
 @Entity(tableName = "turns")
 data class TurnEntity(
