+++ app/src/main/java/com/skgtecnologia/sisem/data/authcards/AuthCardsRepositoryImpl.kt
@@ -3,9 +3,9 @@ package com.skgtecnologia.sisem.data.authcards
 import com.skgtecnologia.sisem.data.auth.cache.AuthCacheDataSource
 import com.skgtecnologia.sisem.data.authcards.cache.AuthCardsCacheDataSource
 import com.skgtecnologia.sisem.data.authcards.remote.AuthCardsRemoteDataSource
-import com.skgtecnologia.sisem.domain.model.screen.ScreenModel
 import com.skgtecnologia.sisem.domain.authcards.AuthCardsRepository
 import com.skgtecnologia.sisem.domain.authcards.model.ConfigModel
+import com.skgtecnologia.sisem.domain.model.screen.ScreenModel
 import javax.inject.Inject
 
 class AuthCardsRepositoryImpl @Inject constructor(
@@ -25,6 +25,6 @@ class AuthCardsRepositoryImpl @Inject constructor(
     override suspend fun getAuthCardsScreen(): ScreenModel =
         authCardsRemoteDataSource.getAuthCardsScreen(
             code = "1", // FIXME: Ambulance code service
-            turnId = authCacheDataSource.retrieveAccessToken()?.turn?.id?.toString() ?: ""
+            turnId = authCacheDataSource.retrieveAccessToken()?.turn?.id?.toString().orEmpty()
         ).getOrThrow()
 }
