+++ app/src/main/java/com/skgtecnologia/sisem/domain/auth/model/AccessTokenModel.kt
@@ -1,14 +1,14 @@
-package com.skgtecnologia.sisem.domain.login.model
+package com.skgtecnologia.sisem.domain.auth.model
 
 data class AccessTokenModel(
+    val userId: Int,
     val accessToken: String,
     val refreshToken: String,
     val tokenType: String,
     val username: String,
     val role: String,
-    val preoperational: PreOperationalModel,
-    val turn: TurnModel,
     val isAdmin: Boolean,
-    val userId: Int,
-    val nameUser: String
+    val nameUser: String,
+    val preoperational: PreOperationalModel,
+    val turn: TurnModel
 )
