+++ app/src/main/java/com/skgtecnologia/sisem/domain/auth/model/TurnModel.kt
@@ -1,4 +1,4 @@
-package com.skgtecnologia.sisem.domain.login.model
+package com.skgtecnologia.sisem.domain.auth.model
 
 data class TurnModel(
     val id: Int,
