+++ app/src/main/java/com/skgtecnologia/sisem/domain/auth/usecases/GetAllAccessTokens.kt
@@ -1,17 +1,17 @@
-package com.skgtecnologia.sisem.domain.login.usecases
+package com.skgtecnologia.sisem.domain.auth.usecases
 
 import androidx.annotation.CheckResult
 import com.skgtecnologia.sisem.commons.extensions.resultOf
 import com.skgtecnologia.sisem.domain.auth.AuthRepository
-import com.skgtecnologia.sisem.domain.login.model.AccessTokenModel
+import com.skgtecnologia.sisem.domain.auth.model.AccessTokenModel
 import javax.inject.Inject
 
-class GetAllAccessToken @Inject constructor(
+class GetAllAccessTokens @Inject constructor(
     private val authRepository: AuthRepository
 ) {
 
     @CheckResult
     suspend operator fun invoke(): Result<List<AccessTokenModel>> = resultOf {
-        authRepository.getAllAccessToken()
+        authRepository.getAllAccessTokens()
     }
 }
