+++ app/src/main/java/com/skgtecnologia/sisem/domain/auth/usecases/Login.kt
@@ -1,9 +1,9 @@
-package com.skgtecnologia.sisem.domain.login.usecases
+package com.skgtecnologia.sisem.domain.auth.usecases
 
 import androidx.annotation.CheckResult
 import com.skgtecnologia.sisem.commons.extensions.resultOf
 import com.skgtecnologia.sisem.domain.auth.AuthRepository
-import com.skgtecnologia.sisem.domain.login.model.AccessTokenModel
+import com.skgtecnologia.sisem.domain.auth.model.AccessTokenModel
 import javax.inject.Inject
 
 class Login @Inject constructor(
