+++ app/src/main/java/com/skgtecnologia/sisem/ui/menu/MenuUiState.kt
@@ -1,9 +1,9 @@
 package com.skgtecnologia.sisem.ui.menu
 
-import com.skgtecnologia.sisem.domain.login.model.AccessTokenModel
+import com.skgtecnologia.sisem.domain.auth.model.AccessTokenModel
 
 data class MenuUiState(
-    val isLoading: Boolean = false,
+    val accessTokenModelList: List<AccessTokenModel>? = null,
     val isLogout: Boolean = false,
-    val accessTokenModelList: List<AccessTokenModel>? = null
+    val isLoading: Boolean = false
 )
